Move Inventory to dedicated source for slightly better code organization
